+++ src/Cli/ComponentClassTable.go
@@ -150,7 +150,7 @@ func (v *ComponentTableClass) Init() {
 				strconv.Itoa(int(*class.Stmts.Analyze.Volume.Loc)),
 				strconv.Itoa(int(*class.Stmts.Analyze.Complexity.Cyclomatic)),
 				strconv.Itoa(int(*class.Stmts.Analyze.Volume.HalsteadLength)),
-				fmt.Sprintf("%.2f", ToFixed(float64(*class.Stmts.Analyze.Volume.HalsteadVolume), 2)),
+				fmt.Sprintf("%.2f", ToFixed(*class.Stmts.Analyze.Volume.HalsteadVolume, 2)),
 				DecorateMaintainabilityIndex(int(*class.Stmts.Analyze.Maintainability.MaintainabilityIndex), class.Stmts.Analyze),
 			})
 		}
@@ -171,7 +171,7 @@ func (v *ComponentTableClass) Init() {
 					strconv.Itoa(int(*class.Stmts.Analyze.Volume.Loc)),
 					strconv.Itoa(int(*class.Stmts.Analyze.Complexity.Cyclomatic)),
 					strconv.Itoa(int(*class.Stmts.Analyze.Volume.HalsteadLength)),
-					fmt.Sprintf("%.2f", ToFixed(float64(*class.Stmts.Analyze.Volume.HalsteadVolume), 2)),
+					fmt.Sprintf("%.2f", ToFixed(*class.Stmts.Analyze.Volume.HalsteadVolume, 2)),
 					DecorateMaintainabilityIndex(int(*class.Stmts.Analyze.Maintainability.MaintainabilityIndex), class.Stmts.Analyze),
 				})
 			}
